LPS-146537 Rename the custom-select directory to CustomSelect
