+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/CustomSelect/CustomSelect.scss
